test-requirements.txt
google-auth>=1.12.0
google-auth-oauthlib>=0.4.1
vcrpy
pytest==7.4.0
pytest-vcr
urllib3==2.0.3
pandas
gspread-dataframe